Setup mesh motion
The precursor domain will be shifted by `inletElevation` vertically and a
python script, `constant/translateBoundaryData.py`, will modify
`constant/boundaryData/$inflowDir/points` so that the the boundary data
generated by `$precursorDir/postProcessing/makeBoundaryDataFiles.$inflowDir.sh`
will coincide with the new mesh.

These input files assume that the surface geometry (provided in
`constant/triSurface/terrain.stl`) is smoothed and flattened at the inlet
boundary, and has a (constant) minimum elevation equal to `inletElevation`.
This is _NOT_ automatically done in `runscript.preprocess.terrain`.

This simulation approach assumes that the boundaryData inlet will be coincident
with the terrain-resolving mesh inlet. This forgoes the assumption of one-to-one
mapping between the original and deformed mesh, and also eliminates the need to
rescale the velocity field up to the PBL height.